Updated JSONs for icon, dirt and glass textures
Updated JSONS for icon, glass, and dirt textures to the latest spec.  Used dates from the zipped "big download" date-modified timestamps.

Leaving grass and course-dirt alone for now, as there seems to be discussion on their implementation so stuff may change.  There's also some ctm-glass textures floating around in the big zip but for now just updating the images that are there.